# api-docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: CLM-EXT-Learning-objects API
  version: 1.0.0
  description: API endpoints the clm-ext-learning_objects module offers
servers:
- url: "{scheme}://{hostname}:{port}{path}"
  description: The production API server
  variables:
    hostname:
      default: localhost
    port:
      default: '5002'
    path:
      default: ''
    scheme:
      default: http
security:
- bearerAuth: []
paths:
...
  "/learningObjects/mgmt/groups/{id}/enrollments/{loId}":
    post:
      tags:
      - mgmt-groups
      - mgmt-los
      summary: 'Enroll a group to a learning object [Minimum role: ''Instructor'']'
      description: Enrolls a specific group to a particular learning object using
        the respective identifiers.
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: loId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successfully enrolled the group to the learning object.
  "/learningObjects/mgmt/groups/{id}/enrollments/{relationId}":
    delete:
      tags:
      - mgmt-groups
      - mgmt-los
      summary: 'Delete Enrollment [Minimum role: ''Instructor'']'
      description: Removes a group's enrollment to a specific learning object, identified
        by its relationId.
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: relationId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successfully deleted the group's enrollment to the learning
            object.
    patch:
      tags:
      - mgmt-groups
      - mgmt-los
      summary: 'Change the enrollment order within the group [Minimum role: ''Instructor'']'
      description: Adjusts the sequence of a specific enrollment within a group based
        on the given relationId.
      parameters:
      - "$ref": "#/components/parameters/accessToken"
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: relationId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successfully modified the order of the enrollment within the
            group.
...
components:
...
  parameters:
    accessToken:
      name: x-access-token
      in: header
      description: The access token
      required: true
      example: exampleAccessToken
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    refreshAuth:
      type: apiKey
      in: header
      name: x-refresh-token
...